refactor: util 패키지를 view 패키지 내부로 이동
