+++ src/main/java/lotto/view/util/Console.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.view.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
