+++ src/main/java/lotto/view/util/InputParser.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.view.util;
 
 import java.util.Arrays;
 import java.util.List;
